requirements.txt
pandas
openpyxl
numpy
psycopg2-binary
beautifulsoup4
langchain
transformers
bitsandbytes
accelerate
scipy
sentencepiece
torch
llama-cpp-python
tabulate
llama-index
openai
pinecone-client[grpc]
tiktoken
datasets
python-dotenv
faiss-cpu
pyyaml
pypdf
sentence-transformers
